+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/BillboardRenderPass.cpp
@@ -182,8 +182,8 @@ void FBillboardRenderPass::Render(const std::shared_ptr<FEditorViewportClient>& 
         else if (UTextComponent* TextComp = Cast<UTextComponent>(BillboardComp))
         {
             FBufferInfo Buffers;
-            float Height = TextComp->Texture->Height;
-            float Width = TextComp->Texture->Width;
+            float Height = static_cast<float>(TextComp->Texture->Height);
+            float Width = static_cast<float>(TextComp->Texture->Width);
             BufferManager->CreateUnicodeTextBuffer(TextComp->GetText(), Buffers, Width, Height, TextComp->GetColumnCount(), TextComp->GetRowCount());
 
             UpdateSubUVConstant(FVector2D(), FVector2D(1, 1));
